wip: add mocks and unit tests
